Fix type numbers
Fixed type number for `fcannounce` and `fclogin`. 

Added `fcrank` too.

Removing the `sign` type because I haven't validated it and it probably doesn't work. WTB shared library with all chat types so I don't have to try to maintain this myself. And because complex/grouped chat types are hell and there isn't a ghood interface to selectively pick subsets of types and adding a custom definition for each one sucks.